requirements.txt
qrcode==7.4.2
pillow==10.2.0
opencv-python==4.9.0.80
pyzbar==0.1.9
PyQt5==5.15.10 